Moved JoinCluster to cluster package and allow to join cluster after labels are set.
